audit_rag/pyproject.toml
[tool.poetry]
name = "audit-rag"
version = "0.1.0"
description = "RAG multi-modal pour l'analyse de rapports d'audit"
authors = ["Votre Nom <vous@example.com>"]
 
[tool.poetry.dependencies]
python          = "^3.10"
langchain       = "^0.2"
langchain-openai= "^0.1"
langchain-community = "^0.2"
langchain-chroma = "^0.1"
langchain-huggingface = "^0.0"
chromadb        = "^0.5"
openai          = "^1.30"
pdfplumber      = "^0.11"
PyMuPDF         = "^1.24"
pdf2image       = "^1.17"
pillow          = "^10.0"
transformers    = "^4.40"
sentence-transformers = "^3.0"
torch           = "^2.3"
ragas           = "^0.1"
fastapi         = "^0.111"
uvicorn         = {extras = ["standard"], version = "^0.29"}
pydantic        = "^2.7"
pydantic-settings = "^2.2"
python-dotenv   = "^1.0"
pandas          = "^2.2"
tiktoken        = "^0.7"
tqdm            = "^4.66"
rich            = "^13.7"
datasets        = "^2.19"
 
[tool.poetry.group.dev.dependencies]
pytest          = "^8.0"
pytest-asyncio  = "^0.23"
pytest-cov      = "^5.0"
mypy            = "^1.10"
ruff            = "^0.4"
black           = "^24.4"
httpx           = "^0.27"
 
[build-system]
requires      = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
 
[tool.ruff]
line-length = 100
select      = ["E", "F", "I", "N", "UP"]
 
[tool.mypy]
python_version = "3.10"
strict         = true